raw-storage/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pydantic==2.6.4
